refactor: extract the node serialization pipeline into SerializesWireNode
