+++ src/Core/Components/Component.php
@@ -4,16 +4,10 @@ declare(strict_types=1);
 namespace Lattice\Lattice\Core\Components;
 
 use JsonSerializable;
-use Lattice\Lattice\Attributes\AsComponent;
-use Lattice\Lattice\Attributes\SerializationHook;
 use Lattice\Lattice\Core\Components\Concerns\HasDataBindings;
-use Lattice\Lattice\Core\Components\Concerns\SerializesToWire;
+use Lattice\Lattice\Core\Components\Concerns\SerializesWireNode;
 use Lattice\Lattice\Core\Concerns\GatesRendering;
 use Lattice\Lattice\Core\Contracts\Renderable;
-use Lattice\Lattice\Support\Wire;
-use ReflectionMethod;
-use Spatie\Attributes\Attributes;
-use Spatie\Attributes\AttributeTarget;
 
 /**
  * @phpstan-consistent-constructor
@@ -22,20 +16,15 @@ abstract class Component implements JsonSerializable, Renderable
 {
     use GatesRendering;
     use HasDataBindings;
-    use SerializesToWire;
-
-    /**
-     * @var array<class-string, list<string>>
-     */
-    private static array $serializationHookCache = [];
+    use SerializesWireNode;
 
     protected bool $hideWhenCollapsed = false;
 
     public function __construct(protected ?string $key = null) {}
 
-    protected function type(): string
+    protected function wireKey(): ?string
     {
-        return AsComponent::typeForClass(static::class);
+        return $this->key;
     }
 
     public function key(string $key): static
@@ -72,45 +61,6 @@ abstract class Component implements JsonSerializable, Renderable
     }
 
     /**
-     * @return array<string, mixed>
-     */
-    public function jsonSerialize(): array
-    {
-        return array_reduce(
-            $this->serializationHooks(),
-            fn (array $data, string $hook): array => $this->{$hook}($data),
-            [],
-        );
-    }
-
-    /**
-     * @param  array<string, mixed>  $data
-     * @return array<string, mixed>
-     */
-    #[SerializationHook(priority: 100)]
-    protected function serialiseBase(array $data): array
-    {
-        return [
-            ...$data,
-            'type' => $this->type(),
-            'key' => $this->key,
-        ];
-    }
-
-    /**
-     * @param  array<string, mixed>  $data
-     * @return array<string, mixed>
-     */
-    #[SerializationHook(priority: 200)]
-    protected function serialiseProps(array $data): array
-    {
-        return [
-            ...$data,
-            'props' => Wire::map($this->decorateProps($this->wireProps())),
-        ];
-    }
-
-    /**
      * @param  array<string, mixed>  $props
      * @return array<string, mixed>
      */
@@ -123,33 +73,5 @@ abstract class Component implements JsonSerializable, Renderable
         }
 
         return $props;
-    }
-
-    /**
-     * @param  array<string, mixed>  $data
-     * @return array<string, mixed>
-     */
-    #[SerializationHook(priority: 10000)]
-    protected function filterEmptyValues(array $data): array
-    {
-        return array_filter(
-            $data,
-            fn (mixed $value, string $key): bool => $key === 'props' || ($value !== null && $value !== []),
-            ARRAY_FILTER_USE_BOTH,
-        );
-    }
-
-    /**
-     * @return list<string>
-     */
-    private function serializationHooks(): array
-    {
-        return self::$serializationHookCache[static::class] ??= collect(Attributes::find($this, SerializationHook::class))
-            ->filter(fn (AttributeTarget $target): bool => $target->attribute instanceof SerializationHook && $target->target instanceof ReflectionMethod)
-            ->filter(fn (AttributeTarget $target): bool => ! $target->target->isPrivate())
-            ->sortBy(fn (AttributeTarget $target): array => [$target->attribute->priority, $target->name])
-            ->map(fn (AttributeTarget $target): string => $target->name)
-            ->values()
-            ->all();
     }
 }
