+++ src/Core/Components/Concerns/SerializesToWire.php
@@ -18,11 +18,6 @@ use ReflectionProperty;
 trait SerializesToWire
 {
     /**
-     * @var array<class-string, list<ReflectionProperty>>
-     */
-    private static array $wirePropertyCache = [];
-
-    /**
      * Reflects the public typed properties (including inherited and trait
      * properties) into the full wire shape: every initialized prop is emitted,
      * keeping null and empty-array values so the payload mirrors the generated
@@ -69,7 +64,10 @@ trait SerializesToWire
      */
     private static function wireProperties(string $class): array
     {
-        return self::$wirePropertyCache[$class] ??= array_values(array_filter(
+        /** @var array<class-string, list<ReflectionProperty>> $cache */
+        static $cache = [];
+
+        return $cache[$class] ??= array_values(array_filter(
             new ReflectionClass($class)->getProperties(ReflectionProperty::IS_PUBLIC),
             static fn (ReflectionProperty $property): bool => ! $property->isStatic(),
         ));
